+++ AnimationTransformer.py
@@ -4,7 +4,7 @@ import time
 import torch
 import torch.nn as nn
 
-import dataset_helper
+import prototype_dataset_helper
 
 
 class AnimationTransformer(nn.Module):
